Extract compactor craft data into CompactorCraftApi
