+++ src/main/java/at/hannibal2/skyhanni/features/inventory/CompactorGfSKeybind.kt
@@ -1,78 +1,22 @@
 package at.hannibal2.skyhanni.features.inventory
 
 import at.hannibal2.skyhanni.SkyHanniMod
-import at.hannibal2.skyhanni.SkyHanniMod.launch
+import at.hannibal2.skyhanni.api.CompactorCraftApi
 import at.hannibal2.skyhanni.api.GetFromSackApi
 import at.hannibal2.skyhanni.api.event.HandleEvent
-import at.hannibal2.skyhanni.config.commands.CommandCategory
-import at.hannibal2.skyhanni.config.commands.CommandRegistrationEvent
-import at.hannibal2.skyhanni.data.SackApi
 import at.hannibal2.skyhanni.events.minecraft.KeyDownEvent
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.ChatUtils
-import at.hannibal2.skyhanni.utils.DelayedRun
-import at.hannibal2.skyhanni.utils.InventoryUtils.getAmountInInventory
-import at.hannibal2.skyhanni.utils.ItemUtils
 import at.hannibal2.skyhanni.utils.ItemUtils.getInternalNameOrNull
 import at.hannibal2.skyhanni.utils.ItemUtils.repoItemName
 import at.hannibal2.skyhanni.utils.NeuInternalName
-import at.hannibal2.skyhanni.utils.NeuItems
 import at.hannibal2.skyhanni.utils.compat.slotUnderCursor
-import at.hannibal2.skyhanni.utils.coroutines.CoroutineSettings
 import net.minecraft.world.entity.player.Inventory
-import kotlin.time.Duration
 
 @SkyHanniModule
 object CompactorGfSKeybind {
 
     private val config get() = SkyHanniMod.feature.inventory.gfs
-
-    /** Null until the first repo reload build finished. Replaced as a whole, never field by field. */
-    @Volatile
-    private var lookup: CraftLookup? = null
-
-    private val repoReloadCoroutine = CoroutineSettings("compactor gfs single step crafts", Duration.INFINITE)
-
-    @HandleEvent
-    private fun onNeuRepoReload() {
-        // The sack item list is only rebuilt after this event, so the build has to wait for it.
-        DelayedRun.runNextTickEnd("compactor gfs single step crafts") {
-            repoReloadCoroutine.launch {
-                buildUpgrades()
-            }
-        }
-    }
-
-    private fun buildUpgrades() {
-        val picked = mutableMapOf<NeuInternalName, Upgrade>()
-        val unclear = mutableMapOf<NeuInternalName, List<Upgrade>>()
-        for ((base, results) in createCandidates()) {
-            // Assumed, not verified: the compactor takes the largest craft. Ice makes Packed Ice at 9
-            // and Enchanted Ice at 160, and 160 is what the compactor produces.
-            val highest = results.maxOf { it.baseAmount }
-            val tied = results.filter { it.baseAmount == highest }
-            tied.singleOrNull()?.let { picked[base] = it } ?: run { unclear[base] = tied }
-        }
-        lookup = CraftLookup(picked, unclear)
-    }
-
-    private fun createCandidates(): Map<NeuInternalName, List<Upgrade>> = NeuItems.allNeuRepoInternalNames()
-        // The compactor puts its result into a sack, so anything that has no sack cannot be one.
-        .filter { it.asString() in SackApi.sackListInternalNames }
-        .flatMap { result -> upgradesFor(result) }
-        .groupBy { it.base }
-
-    private fun upgradesFor(result: NeuInternalName): List<Upgrade> = NeuItems.getRecipes(result)
-        .filter { it.isCraftingRecipe() }
-        .mapNotNull { recipe ->
-            val (base, amount) =
-                ItemUtils.neededItems(recipe).entries.singleOrNull() ?: return@mapNotNull null
-            if (amount <= 1) return@mapNotNull null
-            // The base item is what gets pulled with /gfs later, so it needs a sack of its own.
-            if (base.asString() !in SackApi.sackListInternalNames) return@mapNotNull null
-
-            Upgrade(base, amount, result)
-        }
 
     @HandleEvent(onlyOnSkyblock = true)
     private fun onKeyDown(event: KeyDownEvent) {
@@ -87,69 +31,14 @@ object CompactorGfSKeybind {
     private fun grabMissing(internalName: NeuInternalName) {
         // Ignore silently, the previous request for this item is still on its way.
         if (GetFromSackApi.isQueued(internalName) || GetFromSackApi.wasRecentlySent(internalName)) return
-        val lookup = lookup ?: run {
-            ChatUtils.userError("Recipe data is not loaded yet.")
-            return
-        }
         val itemName = internalName.repoItemName
 
-        if (internalName in lookup.ambiguous) {
-            ChatUtils.userError("$itemName §chas more than one craft at the same amount.")
-            return
-        }
-        val upgrade = lookup.upgrades[internalName] ?: run {
-            ChatUtils.userError("$itemName §ccannot be crafted into another item.")
-            return
-        }
-
-        val missing = upgrade.baseAmount - internalName.getAmountInInventory()
-        if (missing <= 0) {
-            ChatUtils.userError("You already have enough $itemName §cfor ${upgrade.result.repoItemName}§c.")
-            return
-        }
-
-        GetFromSackApi.getFromSack(internalName, missing)
-    }
-
-    @HandleEvent
-    private fun onCommandRegistration(event: CommandRegistrationEvent) {
-        event.registerBrigadier("shtestcompactorcrafts") {
-            description = "List all items with more than one single step craft at the same highest amount"
-            category = CommandCategory.DEVELOPER_DEBUG
-            simpleCallback {
-                printAmbiguous()
-            }
+        when (val state = CompactorCraftApi.getCraftState(internalName)) {
+            NotLoaded -> ChatUtils.userError("Recipe data is not loaded yet.")
+            NoCraft -> ChatUtils.userError("$itemName §ccannot be crafted into another item.")
+            is Ambiguous -> ChatUtils.userError("$itemName §chas more than one craft at the same amount.")
+            is Enough -> ChatUtils.userError("You already have enough $itemName §cfor ${state.upgrade.result.repoItemName}§c.")
+            is Missing -> GetFromSackApi.getFromSack(internalName, state.amount)
         }
     }
-
-    private fun printAmbiguous() {
-        val ambiguous = lookup?.ambiguous ?: run {
-            ChatUtils.userError("Recipe data is not loaded yet.")
-            return
-        }
-        if (ambiguous.isEmpty()) {
-            ChatUtils.chat("No ambiguous single step crafts found.")
-            return
-        }
-        val lines = ambiguous.map { (base, results) ->
-            "${base.asString()} -> ${results.joinToString(", ") { "x${it.baseAmount} into ${it.result.asString()}" }}"
-        }.sorted()
-
-        for (line in lines) {
-            ChatUtils.consoleLog("ambiguous single step craft: $line")
-        }
-        ChatUtils.clickToClipboard("Found ${lines.size} ambiguous single step crafts.", lines)
-    }
-
-    /** A single step craft: [baseAmount] of [base] makes one [result]. */
-    private data class Upgrade(val base: NeuInternalName, val baseAmount: Int, val result: NeuInternalName)
-
-    /**
-     * One finished build. [ambiguous] holds base items where two or more crafts tie at the highest
-     * amount, so no single upgrade can be picked.
-     */
-    private data class CraftLookup(
-        val upgrades: Map<NeuInternalName, Upgrade>,
-        val ambiguous: Map<NeuInternalName, List<Upgrade>>,
-    )
 }
